Houdini project file duplication and hot reload support
This is a intermediate/stub check in to support a workaround for hot reload of sorts for HDK development. The goal is to be able to locate a running Houdini instance of this project, and either refresh its asset libraries or close that instance, write/overwrite the .dll file, and then reopen/reload the Houdini .hiplc file.